+++ packages/expo-image/src/web/ImageWrapper.tsx
@@ -1,14 +1,12 @@
-import React, { useEffect, Ref, useMemo, useState } from 'react';
+import React, { useEffect, Ref } from 'react';
 
 import ColorTintFilter, { getTintColorStyle } from './ColorTintFilter';
-import { ImageWrapperProps, OnErrorEvent } from './ImageWrapper.types';
+import { ImageWrapperProps } from './ImageWrapper.types';
 import { getImageWrapperEventHandler } from './getImageWrapperEventHandler';
+import { useHeaders, useImageHashes } from './hooks';
 import { absoluteFilledPosition, getObjectPositionFromContentPositionObject } from './positioning';
 import { SrcSetSource } from './useSourceSelection';
 import { ImageNativeProps, ImageSource } from '../Image.types';
-import { useBlurhash } from '../utils/blurhash/useBlurhash';
-import { isThumbhashString } from '../utils/resolveSources';
-import { thumbHashStringToDataURL } from '../utils/thumbhash/thumbhash';
 
 function getFetchPriorityFromImagePriority(priority: ImageNativeProps['priority'] = 'normal') {
   return priority && ['low', 'high'].includes(priority) ? priority : 'auto';
@@ -22,72 +20,6 @@ function getImgPropsFromSource(source: ImageSource | SrcSetSource | null | undef
     };
   }
   return {};
-}
-
-function useThumbhash(source: ImageSource | null | undefined) {
-  const isThumbhash = isThumbhashString(source?.uri || '');
-  const strippedThumbhashString = source?.uri?.replace(/thumbhash:\//, '') ?? '';
-  const thumbhashSource = useMemo(
-    () => (isThumbhash ? { uri: thumbHashStringToDataURL(strippedThumbhashString) } : null),
-    [strippedThumbhashString, isThumbhash]
-  );
-  return thumbhashSource;
-}
-
-function useImageHashes(source: ImageSource | null | undefined) {
-  const thumbhash = useThumbhash(source);
-  const blurhash = useBlurhash(source);
-  return useMemo(
-    () => ({
-      resolvedSource: blurhash ?? thumbhash ?? source,
-      isImageHash: !!blurhash || !!thumbhash,
-    }),
-    [blurhash, thumbhash]
-  );
-}
-
-function useHeaders(
-  source: ImageSource | null | undefined,
-  cachePolicy: ImageNativeProps['cachePolicy'],
-  onError?: OnErrorEvent[]
-): ImageSource | null | undefined {
-  const [objectURL, setObjectURL] = useState<string | null>(null);
-  useEffect(() => {
-    (async () => {
-      if (!source?.headers || !source.uri) {
-        return;
-      }
-      try {
-        const result = await fetch(source.uri, {
-          headers: source.headers,
-          cache: cachePolicy === 'none' ? 'no-cache' : 'default',
-          redirect: 'follow',
-        });
-        if (!result.ok) {
-          throw new Error(`Failed to fetch image: ${result.status} ${result.statusText}`);
-        }
-        const blob = await result.blob();
-        setObjectURL(URL.createObjectURL(blob));
-      } catch (error) {
-        onError?.forEach((e) => e?.({ source }));
-      }
-    })();
-  }, [source]);
-  useEffect(() => {
-    return () => {
-      if (objectURL) {
-        URL.revokeObjectURL(objectURL);
-      }
-    };
-  }, [objectURL]);
-  if (!source?.headers) {
-    return source;
-  }
-  if (!objectURL) {
-    // Avoid fetching a URL without headers if we have headers
-    return null;
-  }
-  return { ...source, uri: objectURL };
 }
 
 const ImageWrapper = React.forwardRef(
